MI-3740: Code refactor
