+++ src/components/Modal/index.ts
@@ -1,4 +1,2 @@
 export {Modal} from './Modal.component';
 export type {ModalProps} from './Modal';
-export type {ModalFooterProps} from './subComponents/ModalFooter';
-export type {ModalHeaderProps} from './subComponents/ModalHeader';
