+++ src/components/PostCard/index.ts
@@ -1,3 +1,2 @@
 export {PostCard} from './PostCard.component';
 export type {PostCardProps} from './PostCard';
-export type {PostHeaderProps} from './subComponents/PostHeader';
