+++ src/components/index.ts
@@ -34,13 +34,13 @@ export type {IconProps, IconType} from './Icon';
 export type {InputProps} from './Input';
 export type {LinearProgressProps} from './LinearProgress';
 export type {MenuItemProps} from './MenuItem';
-export type {ModalProps, ModalHeaderProps, ModalFooterProps} from './Modal';
+export type {ModalProps} from './Modal';
 export type {NoDataProps} from './NoData';
 export type {OverlayProps} from './Overlay';
 export type {RadioProps} from './Radio';
 export type {SelectProps} from './Select';
 export type {CircularSkeleton, RectangularSkeleton, SkeletonProps} from './Skeleton';
-export type {SpinnerProps, SpinnerContainerProps} from './Spinner';
+export type {SpinnerProps} from './Spinner';
 export type {SwitchProps} from './Switch';
 export type {TabsProps} from './Tabs';
 export type {HighlightTag, LabelTag, ShortcutTag, TagProps} from './Tag';
@@ -49,4 +49,4 @@ export type {TooltipProps} from './Tooltip';
 export type {MultiProgressBarProps, ProgressBarProps} from './MultiProgress';
 export type {CountdownProps} from './Countdown';
 export type {MMSearchProps} from './MMSearch';
-export type {PostCardProps, PostHeaderProps} from './PostCard';
+export type {PostCardProps} from './PostCard';
